Created some of the basic value types and a few methods and functions to go along with them.
A lot of the stuff is a good start, but not guaranteed to be stable yet. I need to try some things out and build some types and functions to figure out what I need and how I can make things interface and still build reasonably efficient code.

Added an enum, union, and class for Functions and Values. Value has some constructors and methods to do things like get types and union members. Fn has some constructors and an apply function to take care of applying different types of functions. I would really like to be able to compile with overloaded apply functions to not have to pass arrays of args, but there will be so many places where I cannot do that because I won't actually know if a symbol is a function or what function it is. I can find other ways to call core functions directly instead of calling them through apply when it I can be sure they are the core function and not an ambigous symbol.

Added a compiler function for applying values to lists of arguments. I also added a core library with a single const value for holding the + function and the underlying function that it holds. I set this up to work with both doubles and ints and created a couple simple tests.

Overall, things build and the simple tests pass. I will need to make some more tests and clean up and expand all the type classes. I will also probably want to put together another test file for running simple benchmarks to try and see what kinds of implementation details or compiler generated code will affect the runtime of certain things the most. Pretty easy for simple stuff to put some different options in big loops and check how fast those loops run.