Added some headers for necessary classes.
Most of the work here was done creating a stream class and testing a possible module for I/O operations that might target the stream class. This is part to think about some things we might want to use for streams and to think about what we might want a module to look like. I wrote a couple quick tests for the module just to comfirm that the general use case was ok, but it does not test the full design.

One major question I have after trying this is what a character is. I did write a char get function, but it had issues and the notion of a character is somewhat ambiguous in C++. I probably can write a get_byte function for now if I want or just wrap i/o functions to work with strings efficiently where we might want to apply operations to each byte. For example, lowering a string or string find. Also, since I will eventually write a read function and the compiler will need one we need to have the ability to get a single byte at a time, or to work with buffered data by getting a line at a time and then working on the string directly. So, it would be nice to know how I am going to do that kind of thing, but it can wait.