Pulled val, fn, and stream types into their file.
I don't like these all being before their classes, especially since they really need some helpers to make viewing them easier. I extracted them to a header and src file with a namespace and operator<< overloads for viewing the types. Any further enums of this type can be added and have the operator and str methods overloaded. The only thing I am not sure of right now is whether the representation should be a display rep that will be used to indicate types to the user at runtime or if it should be for internal use only and other places can setup a rep for the user. For now I tested only a couple variants for each one just to ensure that the general idea works, but left it at that so I can change the rep when I want for now.