Updated stream and fn to use the new type.h classes and for fn to keep its union internal.
I also decided that all the internal classes and functions should use C++ error, or a separate custom error from the lisccp error classes. The reason for this is that I want to handle errors related to library programming errors and user errors differently. Or at least separate where I pay attention to eahc. So errors that are likely a result of poorly compiled code, or mistakes interacting with the library code will be C++ error. Then errors that occur from the user making a mistake will be liscpp errors that will be easier to reason about and will be possible to interact with in liscpp code.:w I might still need to tweak this, but I think there is an important distinction here.