Fixed value to use the type and put the union inside.
This was a lot of changes. The updates to Value hit pretty much every file. I moved the union inside like I did with fn, I also did this for stream. I made all initializers static member functions. I am still not sure this is the absolute right move, but it works for now. The problem is with ambiguous constructors when I use literals.

I don't like the static members creating a value and then updating it in the functions, but have discovered I can make private constructors for the specific types and call them through the functions so there is no ambiguity. This might shave a little off the calls, though it is sort of double the effort. I also think it might be best just to make them functions in a value namespace, make the constructors public, but just call them through the functions in tests etc. Perhaps in compiled code it would not be so bad to be explicit about the arguments, but of course some types have the same underlying type and then an overload will not work, unless it has a switch on a type arg to decide, which makes it better just to call an explicit constructor function.

The main savings here, and the reason for not using inheritance in the first place, is for small allocations. It is possible that the garbage collection allocation would reduce the cost of allocation, since it will allocate a large area and then manage it. So, if ints were malloc'd it would not be as bad as asking for small chunks of memory many times, though it is still better if we can avoid allocating for small things.