Added the rest of the basic tests for the value type.
In order to make this work I had to make some adjustments in a few places. Like implementing a basic list equality etc. I also had to update the typedef for the GcMap and implement comparisson operators for Value. I am not fully happy with the implementation yet, but it is coming. Also comparissons are not tested yet. Not all types make sense with comparissons and I threw errors, but I am not sure this will work when comparing values. The major issue is that in map they will be compared with < and if that throws an error then it would mean the map would essentially be typed by key. So, the comparisson needs to be fixed. Probably, I will just make my own comparator to replace std::less and then that can compare values for the map in a way that makes sense. And if I want to have a full comparisson built into objects I can keep the operator<.

I had some linker errors that I thought were maybe dependency related, bu they were just the usual forgetting to qualify some function definitions with namespaces. While I was working on this I added a panic that will just crash the program. I am still not sure about this or whether places should always throw liscpp errors. I like separating the concerns, and putting in panics makes it my responsibility to properly cover expected issues calling some functions, but also means if I miss it and don't test properly it will crash. I have some panics in value, but I will revisit this before moving on to other things.