Replaced the operator< on value and list with a custom comparator specifically for use with GcMap.
The reason for this is that I am not sure it makes sense to have a < operator that compares values of differing types. To that point the pervious one threw errors if types were not equal, which will not work for a map compare. It also would not really make sense if we do allow sorting maps with multiple types in them.

I created a comparator that can compare all types and if two types are different has a set priority for which will always be lesser. This will allow map to have consistend comparissons and give map order consistency. It does use pointer comparisson for pointer stored values, except for string, symbol, and keyword. The goal is to keep the map comparissons as fast as possible while being consistent. The equality will be deep, but if things are not equal where we put them in the map is not as important. So, it will be an ordered map, but not necessarily ordered for each type, though currently all the primitive types, and string based types will be ordered properly.