Added core namespace with a few initial functions and some tests for one of them.
Added the arithmetic functions. Put in some tests for a + function and implemented it. The others still just return nil. I also forgot to test + for floats or a combination of ints and floats.

I also adjusted the namespacing of the operator overloads to be in the liscpp namespace. I had them outside before because I was not using namespaces for everything and it would only work if they were outside. Now, at least for gtest, it seems better to keep them inside the namespace their types are in. This way I get them to print properly when using EXPECT_EQ in gtest with values.